refactor: move octicons build output to dist
Place the @primer/octicons JavaScript, styles, data, and SVG assets under dist while preserving the legacy build subpath export.

Copilot-Session: ec9f0dbd-ddc0-4fbc-86af-1d73d9b45dbc
