+++ .changeset/modern-pandas-smile.md
@@ -2,4 +2,4 @@
 '@primer/octicons': minor
 ---
 
-Add an ES module entry point while retaining CommonJS support for `@primer/octicons`.
+Add an ES module entry point while retaining CommonJS support for `@primer/octicons`, with published files organized under `dist`.
