+++ lib/octicons_node/.gitignore
@@ -1,2 +1,1 @@
-/index.js
-/index.cjs
+/dist
