+++ turbo.json
@@ -8,7 +8,7 @@
     },
     "build": {
       "dependsOn": ["//#build", "^build"],
-      "outputs": ["dist/**", "build/**", "src/__generated__/**", "index.js", ".next/**"]
+      "outputs": ["dist/**", "build/**", "src/__generated__/**", ".next/**"]
     },
     "//#lint": {
       "dependsOn": ["@primer/octicons#build", "@primer/octicons-react#build", "@primer/styled-octicons#build"],
